Added temp addon items for pushing non-addon positions
If the item we're adding is not an addon, pretix will refuse to add it as addon to another item. However, it won't complain if we update an already existing addon. What we do here is push a temp addon to the item, and then swap its item id with the item we actually want to push.

The temp item must be an addon to the original item, however by placing it in an addon category and by keeping it hidden (not visible) except for vouchers, the category won't appear at all in the add-on choice for the users