Fixed failed login attempts count
After analyzing the code of the LoginUseCase I noticed that the @\Transactional was useless. If we have a crash in the middle of the execution, we may endup with:
- Attempts reset to 0, but no new session created
- Before point + old sessions cleaned, but still no new session created
Both when the user has successfully logged in.
This is not in any possible way an invalid DB state, so I'm fine removing the Transactional.